+++ packages/contract-state-stream/__tests__/node_env.js
@@ -3,7 +3,7 @@
  */
 const createNewBlock$ = require("@drizzle-utils/new-block-stream");
 const { take, finalize, tap, toArray } = require("rxjs/operators");
-const initTestChain = require("@drizzle-utils/test-ganache");
+const initTestChain = require("@drizzle-utils/test-chain");
 
 const createContractState$ = require("../index");
