+++ packages/test-chain/README.md
@@ -1,4 +1,4 @@
-# @drizzle-utils/test-ganache
+# @drizzle-utils/test-chain
 
 Utils for spawning a test blockchain to run automated tests on.
