+++ packages/test-chain/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@drizzle-utils/test-ganache",
+  "name": "@drizzle-utils/test-chain",
   "version": "0.1.6-alpha.0",
   "main": "index.js",
   "license": "MIT",
